+++ CHANGELOG.md
@@ -17,6 +17,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Required JDK to 17
 - Microphone mute to rely on `AudioDeviceModule.setMicrophoneMute(Boolean)` instead of
   `AudioManager.setMicrophoneMute(Boolean)`
+- Noise suppressor and acoustic echo canceler to use software implementation
 
 ## [0.16.0] - 2024-10-02